trace behind debug tag
